Separate MigrationVrmExpression
