+++ Assets/VRM10/Runtime/Migration/MigrationVrmExpression.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4bc0b82a3f20bf5429f1ee36766f5769
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
